+++ common/components/action/SetLocaleAction.php
@@ -8,9 +8,10 @@ namespace common\components\action;
 use yii\base\Action;
 use yii\base\InvalidParamException;
 use Yii;
+use yii\web\Cookie;
 
 /**
- * Class SetLocale
+ * Class SetLocaleAction
  * @package common\components\action
  *
  * Example:
@@ -19,11 +20,11 @@ use Yii;
  *   {
  *       return [
  *           'set-locale'=>[
- *               'class'=>'common\components\actions\SetLocale',
+ *               'class'=>'common\components\actions\SetLocaleAction',
  *               'locales'=>[
  *                   'en-US', 'ru-RU', 'ua-UA'
  *               ],
- *               'localeSessionKey'=>'user.locale',
+ *               'localeCookieName'=>'_locale',
  *               'callback'=>function($action){
  *                   return $this->controller->redirect(/.. some url ../)
  *               }
@@ -32,20 +33,44 @@ use Yii;
  *   }
 */
 
-class SetLocale extends Action
+class SetLocaleAction extends Action
 {
+    /**
+     * @var array List of available locales
+     */
     public $locales;
-    public $localeSessionKey = 'user.locale';
+
+    /**
+     * @var string
+     */
+    public $localeCookieName = '_locale';
+
+    /**
+     * @var integer
+     */
+    public $cookieExpire;
+
+    /**
+     * @var \Closure
+     */
     public $callback;
 
 
-
+    /**
+     * @param $locale
+     * @return mixed|static
+     */
     public function run($locale)
     {
         if(!is_array($this->locales) || !in_array($locale, $this->locales)){
             throw new InvalidParamException('Unacceptable locale');
         }
-        Yii::$app->session->set($this->localeSessionKey, $locale);
+        $cookie = new Cookie([
+            'name' => $this->localeCookieName,
+            'value' => $locale,
+            'expire' => $this->cookieExpire ?: time() + 60 * 60 * 24 * 365,
+        ]);
+        Yii::$app->getResponse()->getCookies()->add($cookie);
         if($this->callback && $this->callback instanceof \Closure){
             return call_user_func_array($this->callback, [
                 $this,