+++ common/components/behaviors/LocaleBehavior.php
@@ -11,7 +11,20 @@ namespace common\components\behaviors;
 use yii\base\Behavior;
 use Yii;
 
+/**
+ * Class LocaleBehavior
+ * @package common\components\behaviors
+ */
 class LocaleBehavior extends Behavior{
+
+    /**
+     * @var string
+     */
+    public $cookieName = '_locale';
+
+    /**
+     * @return array
+     */
     public function events()
     {
         return [
@@ -19,9 +32,12 @@ class LocaleBehavior extends Behavior{
         ];
     }
 
+    /**
+     * Resolve application language by checking user cookies, preferred language and profile settings
+     */
     public function beforeRequest(){
-        if(Yii::$app->session->has('user.locale') && !Yii::$app->session->hasFlash('forceUpdateLocale')){
-            $userLocale = Yii::$app->session->get('user.locale');
+        if(Yii::$app->getRequest()->getCookies()->has($this->cookieName) && !Yii::$app->session->hasFlash('forceUpdateLocale')){
+            $userLocale = Yii::$app->getRequest()->getCookies()->getValue($this->cookieName);
         } else {
             $userLocale = !Yii::$app->user->isGuest
                         && Yii::$app->user->getIdentity()->profile