+++ frontend/controllers/SiteController.php
@@ -24,7 +24,7 @@ class SiteController extends Controller
                 'fixedVerifyCode' => YII_ENV_TEST ? 'testme' : null,
             ],
             'set-locale'=>[
-                'class'=>'common\components\action\SetLocale',
+                'class'=>'common\components\action\SetLocaleAction',
                 'locales'=>array_keys(Yii::$app->params['availableLocales'])
             ]
         ];
@@ -40,9 +40,9 @@ class SiteController extends Controller
         $model = new ContactForm();
         if ($model->load(Yii::$app->request->post()) && $model->validate()) {
             if ($model->sendEmail(Yii::$app->params['adminEmail'])) {
-                Yii::$app->session->setFlash('success', 'Thank you for contacting us. We will respond to you as soon as possible.');
+                Yii::$app->session->setFlash('success', Yii::t('frontend', 'Thank you for contacting us. We will respond to you as soon as possible.'));
             } else {
-                Yii::$app->session->setFlash('error', 'There was an error sending email.');
+                Yii::$app->session->setFlash('error', \Yii::t('frontend', 'There was an error sending email.'));
             }
 
             return $this->refresh();